+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 ## [2.2.0] - 2020-12-02
 ### Added
 - Column selection for resource tables
+- Allowed non-LOINC codes to be entered in Patient selection criteria
 
 ## [2.1.0] - 2020-11-19
 ### Added