+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [2.2.0] - 2020-12-02
+### Added
+- Column selection for resource tables
+- Allowed non-LOINC codes to be entered in Patient selection criteria
+- Added description for each step
+
 ## [2.1.0] - 2020-11-19
 ### Added
 - Reuse selection criteria values from the Patient selection area